requirements.txt
flask
pandas
numpy
Pillow
tensorflow
scikit-learn
opencv-python
matplotlib
plotly
scipy
nltk